backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
passlib
argon2-cffi
python-jose[cryptography]
python-multipart
email-validator
stripe
